[IMP] calendar,hr_holidays,microsoft_calendar: use ACL instead of user_can_edit field for readonly
Since https://github.com/odoo/odoo/pull/112964 and task 3185743 only
attendee can edit an event.

user_can_edit field is used in the UI to put fields in readonly.

Before this commit, this limitation was only cosmetic. It was still
possible to use RPC to update the event. This commit relies on standard
Odoo ACL to prevent event modification. This prevents write calls update an
event when an user is not in the attendees or the organizer.

taskid: 4137628

+++ addons/calendar/models/calendar_attendee.py
@@ -57,6 +57,9 @@ class CalendarAttendee(models.Model):
     @api.model_create_multi
     def create(self, vals_list):
         for values in vals_list:
+            if 'event_id' in values:
+                event = self.env['calendar.event'].browse(values['event_id'])
+                event.check_access('write')
             # by default, if no state is given for the attendee corresponding to the current user
             # that means he's the event organizer so we can set his state to "accepted"
             if 'state' not in values and values.get('partner_id') == self.env.user.partner_id.id:
@@ -68,7 +71,14 @@ class CalendarAttendee(models.Model):
                 values['common_name'] = values.get("common_name")
         return super().create(vals_list)
 
+    def write(self, vals):
+        if 'event_id' in vals:
+            event = self.env['calendar.event'].browse(vals['event_id'])
+            event.check_access('write')
+        return super().write(vals)
+
     def unlink(self):
+        self.event_id.check_access('unlink')
         self._unsubscribe_partner()
         return super().unlink()
